commit a staged change: git commit when you commit, you have to add a message. to avoid complexity, use -m "message" to add message e.g. git commit -m "initial commit"
what if you forgot to add a message? it will open up a VIM editor and allow you to type in the messagecommit a staged change: git commit
when you commit, you have to add a message.
to avoid complexity, use -m "message" to add message
e.g. git commit -m "initial commit"

what if you forgot to add a message? it will open up a VIM editor and allow you to type in the message
initial commit
:wq
git commit
:wq
:wq
:wq